vr_soft_api/_BACKUP_OLD/requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
psycopg2-binary>=2.9.0
python-multipart>=0.0.6
pydantic>=2.0.0
requests>=2.31.0
